feat(cli): expose the Rust CLI through Python
