+++ bamnado-python/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bamnado-python"
-version = "0.8.0"
+version = "0.8.1"
 edition = "2024"
 description = "Python extension wrapper for bamnado"
 license = "MIT OR Apache-2.0"
