+++ bamnado-python/python/bamnado/__main__.py
@@ -1,0 +1,7 @@
+import sys
+
+from .cli import main
+
+
+if __name__ == "__main__":
+    sys.exit(main())
