+++ bamnado-python/python/bamnado/_bamnado.pyi
@@ -127,3 +127,5 @@ def compute_scale_factors(
 def __version__() -> str:
     """Get the version of the bamnado package."""
     ...
+
+def _cli_main(argv: list[str]) -> int: ...
