+++ pyproject.toml
@@ -33,7 +33,7 @@ rustc-args = []
 
 [project]
 name = "bamnado"
-version = "0.8.0"
+version = "0.8.1"
 description = 'Python bindings for BamNado'
 authors = [{name = 'Alastair Smith', email = 'alastair.smith@ndcls.ox.ac.uk'}]
 readme = "README.md"
@@ -59,6 +59,9 @@ dependencies = [
 Homepage = "https://github.com/alsmith151/BamNado"
 Repository = "https://github.com/alsmith151/BamNado"
 Documentation = "https://alsmith151.github.io/BamNado/"
+
+[project.scripts]
+bamnado = "bamnado.cli:main"
 
 [project.optional-dependencies]
 test = ["pytest>=7.0", "pytest-cov>=4.0"]
