+++ scripts/check-versions.sh
@@ -1,0 +1,63 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+if [[ $# -gt 1 ]]; then
+  echo "usage: $0 [expected-version]" >&2
+  exit 2
+fi
+
+if [[ $# -eq 1 ]]; then
+  expected=$1
+else
+  expected=$(sed -nE 's/.*"bamnado"[[:space:]]*:[[:space:]]*"([^"]+)".*/\1/p' .release-please-manifest.json | head -1)
+fi
+
+package_version() {
+  awk '
+    /^\[package\]$/ { in_package = 1; next }
+    /^\[/ { in_package = 0 }
+    in_package && /^version[[:space:]]*=/ { split($0, parts, "\""); print parts[2]; exit }
+  ' "$1"
+}
+
+project_version() {
+  awk '
+    /^\[project\]$/ { in_project = 1; next }
+    /^\[/ { in_project = 0 }
+    in_project && /^version[[:space:]]*=/ { split($0, parts, "\""); print parts[2]; exit }
+  ' pyproject.toml
+}
+
+if [[ -f Cargo.lock ]]; then
+  lock_version=$(awk '
+    /^name = "bamnado-python"$/ { found = 1; next }
+    found && /^version = / { split($0, parts, "\""); print parts[2]; exit }
+  ' Cargo.lock)
+else
+  # Cargo.lock is intentionally ignored in this repository and is generated by
+  # cargo before release artifacts are built.
+  lock_version=""
+fi
+
+locations=("bamnado/Cargo.toml" "bamnado-python/Cargo.toml" "pyproject.toml" "Cargo.lock")
+versions=("$(package_version bamnado/Cargo.toml)" "$(package_version bamnado-python/Cargo.toml)" "$(project_version)" "$lock_version")
+
+failed=0
+for i in "${!locations[@]}"; do
+  location=${locations[$i]}
+  version=${versions[$i]}
+  if [[ "$location" == "Cargo.lock" && -z "$version" ]]; then
+    printf '%s: not present (generated by cargo)\n' "$location"
+    continue
+  fi
+  printf '%s: %s\n' "$location" "$version"
+  if [[ "$version" != "$expected" ]]; then
+    failed=1
+  fi
+done
+
+if [[ $failed -ne 0 ]]; then
+  echo "Version mismatch: expected $expected" >&2
+  exit 1
+fi
+echo "All versions are consistent: $expected"
